This screen is the last screen of Create Profile, and all the states will be stored here and when I click on button finish I push every thing to Firebase
So, Firstly I created 3 states for this screen and then I created total - this on is a container state of 3 screens.
Secondly after I handled 3 states of this screen, I created a function and I put it in a button, so when we click on the button, the function will be called the dispatch.
The dispatch will call the action of CreateProfilePPU