This screen I handle the state of select, input text and finally and very hard thing, that's image preview
So, select state,input text is ok.
But the image, we want that have a preview image, so will have imageHandler function to read the image and convert image to preview image, not only just a file have the name.
And in this screen I have to take the state of screen registration and pass it to screen finish when I click the button next, I will pass the state of the image to screen finish and the last screen can push the image to the storage of firebase